backend/requirements.txt
fastapi>=0.111.0
psycopg2-binary>=2.9.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
alembic>=1.13.0
apscheduler>=3.10.0
pydantic>=2.0.0
httpx>=0.27.0
pandas>=2.2.0
yfinance>=0.2.40
python-dotenv>=1.0.0
pytz>=2024.1
python-multipart>=0.0.9
playwright>=1.45.0
pytest>=8.0.0
pytest-mock>=3.14.0
